Named session back to db, purged endpoints and renamed dataloaders
