+++ DashAI/back/routers/datasets.py
@@ -6,13 +6,13 @@ from fastapi.encoders import jsonable_encoder
 from fastapi.exceptions import HTTPException
 from sqlalchemy import exc
 
-from DashAI.back.database import models, session
+from DashAI.back.database import db, models
 from DashAI.back.dataloaders.classes.csv_dataloader import CSVDataLoader
 
 # from Dataloaders.classes.audioDataLoader import AudioDataLoader
 # from Dataloaders.classes.csvDataLoader import CSVDataLoader
 # from Dataloaders.classes.imageDataLoader import ImageDataLoader
-from DashAI.back.dataloaders.classes.dataload_pydantic import DatasetParams
+from DashAI.back.dataloaders.classes.dataloader_params import DatasetParams
 from DashAI.back.models.classes.getters import get_model_params_from_task
 from DashAI.back.routers.session_class import session_info
 from DashAI.back.tasks.task import Task
@@ -39,23 +39,20 @@ async def get_dataset():
     """
     Returns all the available datasets in the DB.
     """
-
     available_datasets = {}
-
-    for db_dataset in session.session.query(models.Dataset).order_by(models.Dataset.id):
-        act_dataset = {
-            "dataset_id": db_dataset.id,
-            "dataset_name": db_dataset.name,
-            "dataset_task_name": db_dataset.task_name,
-            "dataset_path": db_dataset.path,
-        }
-        available_datasets[db_dataset.id] = act_dataset
-
+    try:
+        for db_dataset in db.session.query(models.Dataset):
+            available_datasets[db_dataset.id] = {
+                "dataset_name": db_dataset.name,
+                "dataset_task_name": db_dataset.task_name,
+                "dataset_path": db_dataset.path,
+            }
+    except exc.SQLAlchemyError:
+        return {"message": "Couldn't connect with DB."}
     return available_datasets
 
 
-# @router.post("/dataset/") TODO Change this route to this
-@router.post("/dataset/upload/")
+@router.post("/dataset/")
 async def upload_dataset(
     params: str = Form(), url: str = Form(None), file: UploadFile = File(None)
 ):
@@ -64,7 +61,8 @@ async def upload_dataset(
     Returns a list of available models for the dataset's task.
     """
     params = parse_params(params)
-    # TODO Globals call is removed on the next iteration of dataloaders branch
+    # TODO Multiple dataloaders, first ideas contain a dict with the structure:
+    # {"CSVDataLoader": CSVDataLoader(), "JSONDataLoader": JSONDataLoader(), ...}
     dataloader = CSVDataLoader()
     folder_path = f"DashAI/back/user_datasets/{params.dataset_name}"
     try:
@@ -90,9 +88,9 @@ async def upload_dataset(
     try:
         folder_path = os.path.realpath(folder_path)
         db_dataset = models.Dataset(params.dataset_name, params.task_name, folder_path)
-        session.session.add(db_dataset)
-        session.session.flush()
-        session.session.commit()
+        db.session.add(db_dataset)
+        db.session.flush()
+        db.session.commit()
 
         # TODO remove this
         session_info.dataset = dataset
@@ -102,14 +100,11 @@ async def upload_dataset(
         return get_model_params_from_task(params.task_name)
     except exc.SQLAlchemyError:
         return {"message": "Couldn't connect with DB."}
+    
+@router.delete("/dataset/")
+async def delete_dataset():
+    return {"message": "To be implemented"}
 
-
-@router.get("/dataset/task_name/{session_id}")
-async def get_task_name():
-    """
-    Returns the task_name associated with the experiment of id session_id.
-    """
-    try:
-        return session_info.task_name
-    except AttributeError:
-        return {"message": "There was a problem obtaining the dataset's task"}
+@router.put("/dataset/")
+async def insert_dataset():
+    return {"message": "To be implemented"}
